working with long double in hits rpm

+++ hits_cpp/hits.cpp
@@ -15,7 +15,7 @@
 
 Memes loadMemes(string memePath, int limit, bool verbose);
 void loadCutoffs(string cutoffsPath, Memes& memes, int limit, bool verbose);
-SequencesMap loadSequences(string faaPath, int& numSequences, SequencesRpmMap& seeuncesRpm, bool useRpmFaaScanning, float rpmFactorValue, bool verbose);
+SequencesMap loadSequences(string faaPath, int& numSequences, SequencesRpmMap& seeuncesRpm, bool useRpmFaaScanning, long double rpmFactorValue, bool verbose);
 
 // TODO support Repeats_
 // TODO move createShuffles, isHit and getHits?
@@ -280,7 +280,7 @@ int main(int argc, char *argv[])
         ("useFactor", "To multiply by factor hits for normalization", cxxopts::value<bool>()->default_value("false"))
         ("sequenceHitMotifPath", "Path for results of sequence that had hit with motif", cxxopts::value<string>()->default_value(""))
         ("useRpmFaaScanning", "Performance of scanning script with rpm faa file", cxxopts::value<bool>()->default_value("false"))
-        ("rpmFactorValue", "RPM factor value of the sample", cxxopts::value<float>()->default_value("1.0"))
+        ("rpmFactorValue", "RPM factor value of the sample", cxxopts::value<long double>()->default_value("1.0"))
         ("v,verbose", "Verbose output", cxxopts::value<bool>()->default_value("false"));
     auto result = options.parse(argc, argv);
 
@@ -296,7 +296,7 @@ int main(int argc, char *argv[])
     auto useFactor = result["useFactor"].as<bool>();
     auto sequenceHitMotifPath = result["sequenceHitMotifPath"].as<string>();
     auto useRpmFaaScanning = result["useRpmFaaScanning"].as<bool>();
-    auto rpmFactorValue = result["rpmFactorValue"].as<float>();
+    auto rpmFactorValue = result["rpmFactorValue"].as<long double>();
     auto isVerbose = result["verbose"].as<bool>();
 
     auto begin = chrono::steady_clock::now();
